The scheduler is complicated (#3)
* The scheduler is complicated

After looking through the documentation, and then looking at the code.. it seems like the best way to avoid long running jobs is to change the disable-window-seconds option in the config files. It's set to use dashes rather than underscores for whatever reason, and there are some mismatches between task and cfg names..
`retry_count == disable_failures`

* change disable failures to 10

This seems like a lot, but it's a good starting point for any individual process. This will be overridden in the futures and custom_targeting classes though.